Task 4-To-Do List Manager
This code provides a console-based interface for managing a to-do list. Let's walk through the functionalities:

Add a Task: The user can enter a task description, which is then added to the list of tasks. Each task is represented by a Task structure, containing a description string and a completed flag.

View Tasks: This option displays all the tasks in the to-do list. Each task is numbered, and if a task is marked as completed, it is displayed with a [X] symbol. Otherwise, it is shown as [].

Remove a Task: Users can specify the task number they want to remove. If the input is valid, the corresponding task is deleted from the list.

Mark Task as Completed: Users can mark a task as completed by providing its task number. Once marked, the task's completed flag is set to true.

Exit: This option allows users to exit the program.

The code uses a vector named tasks to store all the tasks. The main loop continuously prompts the user for their choice and performs the corresponding action based on the input.